requirements.txt
aiogram==2.12.1
aiohttp==3.7.4.post0
aioredis==1.3.1
appdirs==1.4.4
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
black==21.4b0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
hiredis==2.0.0
idna==3.1
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.8.1
py-dotenv==0.1
python-dotenv==0.17.0
pytz==2021.1
regex==2021.4.4
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
yarl==1.6.3
